Fix colormap delete (#155)
* Updates code to use updated vcs-js.

* Makes delete work with vcs.

+++ frontend/src/js/components/modals/ColormapEditor/ColormapWidget.jsx
@@ -105,27 +105,39 @@ class ColormapWidget extends Component {
     }
 
     handleDeleteColormap(){
+        // TODO: If i use a colormap then delete it what happens?
         let nameToDelete = this.state.selectedColormapName
-        let colormapNames = Object.keys(this.props.colormaps).sort(function (a, b) {
-            return a.toLowerCase().localeCompare(b.toLowerCase())
-        })
-        let index = colormapNames.indexOf(nameToDelete)
-        if(index == colormapNames.length - 1){ // if the colormap to delete is the last in the list
-            index = colormapNames.length - 2; // select the colormap before it
+        if(nameToDelete !== "default"){
+            try{
+                if(vcs){ // eslint-disable-line no-undef
+                    vcs.deleteColormap(nameToDelete).then((data)=>{console.log(data)}) // eslint-disable-line no-undef
+                }
+            }
+            catch(e){
+                if(e instanceof ReferenceError){
+                    console.warn("VCS is not defined. Is the VCS Server running?")
+                }
+            }
+            let colormapNames = Object.keys(this.props.colormaps).sort(function (a, b) {
+                return a.toLowerCase().localeCompare(b.toLowerCase())
+            })
+            let index = colormapNames.indexOf(nameToDelete)
+            if(index == colormapNames.length - 1){ // if the colormap to delete is the last in the list
+                index = colormapNames.length - 2; // select the colormap before it
+            }
+            else{
+                index++ // else, select the colormap below it
+            }
+            let name = colormapNames[index]
+            let currentColormap = _.map(this.props.colormaps[name], _.clone())
+            this.props.deleteColormap(nameToDelete)
+            setTimeout(()=>{
+                this.setState({
+                    selectedColormapName: name,
+                    currentColormap: currentColormap,
+                })  
+            }, 0)
         }
-        else{
-            index++ // else, select the colormap below it
-        }
-        let name = colormapNames[index]
-        let currentColormap = _.map(this.props.colormaps[name], _.clone())
-        this.props.deleteColormap(nameToDelete)
-        setTimeout(()=>{
-            this.setState({
-                selectedColormapName: name,
-                currentColormap: currentColormap,
-            })  
-        }, 0)
-        
     }
 
     blendColors(){
@@ -171,7 +183,7 @@ class ColormapWidget extends Component {
 
         /* eslint-disable no-undef */ 
         if(vcs){
-            vcs.colormapnames().then((names) => {
+            vcs.getcolormapnames().then((names) => {
                 if(names.indexOf(colormap_name) >= 0){
                     vcs.setcolormap(colormap_name, self.state.currentColormap).then(() => { // save colormap in vcs
                         self.props.saveColormap(colormap_name, self.state.currentColormap) // save to the frontend state
